Added authentication to update and delete goal

+++ Backend/Controllers/goalController.js
@@ -2,12 +2,13 @@ const asyncHandler = require("express-async-handler");
 
 // Loaded with Mongoose methods.
 const Goal = require("../Models/goalModel");
+const User = require("../Models/UserModel");
 
 // @desc Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({user: req.user.id}); // all of them
+  const goals = await Goal.find({ user: req.user.id }); // all of them
 
   res.status(200).json(goals);
 });
@@ -23,7 +24,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
-    user: req.user.id // connect text to a user
+    user: req.user.id, // connect text to a user
   });
 
   res.status(200).json(goal);
@@ -37,6 +38,19 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not Found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // compare authenticated user matches goal creator
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -54,6 +68,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not Found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // compare authenticated user matches goal creator
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized");
   }
 
   await goal.deleteOne();
